Implemented arm->box movement
When a vehicle leaves an arm it is added to the box.
Fixed a bug in Box where vehicles were removed incorrectly
Junction now calls arms to move vehicles.
Implemented can_enter_box, determining if a car can enter based on the traffic light direction and the direction of other cars in the box.
Added source_lane property to vehicles as it is needed for box movement