Added Pedestrian Crossings
Details of crossings put in the front end are now passed to the back end.

Crossing requests are based on the same exponential distribution used to generate cars. When the interval reaches 0, a crossing request is queued. When the lights next change to red, all lights stay red until the pedestrian crossing period has elapsed.

Added print statements for debugging (currently commented out)