Bus lane backend
Implemented backend side of bus lanes. Busses can use the lane normally, while cars are prevented from merging into them or being created in them.

Bus lanes are created by a parameter to the junction constructor. This creates a bus lane in addition to the given number of car lanes, at the left side of the arm.