Finished left turn lanes
Vehicles can now turn left in left turn lanes when the light is red. If the green light is on the arm to the left, they can always turn, otherwise they can turn if there are no vehicles moving to the same arm. Vehicles cannot turn during pedestrian crossings or while lights change