Traffic light unit tests
Added unit tests for traffic lights without pedestrian crossings, checking that the light cycles as expected.

Changed behaviour when vehicles aren't present, so that if there are no vehicles in any other lane the light does not change. This may introduce a bug with pedestrian crossings which will be fixed when tests are added for them