Pedestrian crossing unit tests
Added unit tests for pedestrian crossings.
Changed traffic light behaviour, so that if there are no cars the lights are all set to red until cars are present. If there is a crossing request during this time the pedestrian crossing starts immediately.

Added a default option for allowed_directions for junctions